clean ebpf folder
